rlp: remove too many bytes error (#37)
In the case that the input to Decode contains more bytes
than is specified by the RLP header, we simply ignore those
bytes instead of returning an error.

This is motivated by the fact that the RLPx protocol will
add padding around an RLP message. To complicate matters,
the padding is random -making it difficult to remove the
padding prior to calling rlp.Decode. Therefore, it seems
fine and good to have Decode simply ignore trailing data.

This commit also removes an unused test helper function.