rlp: remove unused errors. use bint for decoding
bint encoding/decoding: we introduced the bint package after adding the
rlp package. This commits moves the rlp package towards the bint package
which eliminates the need for padding.

unused errors: it seems better to not return errors for things that have
good zero values. Instead, callers can check the length of a string or
byte slice.