abi: add function for parsing event logs. s/at/abit (#48)
Uses the abi encoding functionality to parse emitted events from evm logs.

Indexed logs are easy to parse as they are contained in the topics array.
However, un-indexed logs are abi-encoded bytes in the log's data field.
The data field needs to be decoded with a user-supplied schema.

The `abit` package can be used to define the decoding schema.

Example:

    bxfr, ok := xabi.Match(log, abi.NamedEvent("TransferBatch").
        Indexed("operator", abit.Address).
        Indexed("from", abit.Address).
        Indexed("to", abit.Address).
        UnIndexed("ids", abit.List(abit.Uint256)).
        UnIndexed("values", abit.List(abit.Uint256)),
    )