abi: build type schema from Event/Input structs
The newly added Input.ABIType and abit.Resolve method/function supports
the flow of building Event/Input structs from JSON instead of builder
methods. It is still possible to build the structs by hand, but it is
likely that the majority of use cases will depend on parsing a ABI JSON
file and then using that data to call Match functions. In this case, we
needed a method to recursively build the abit.Type schema from the JSON.