abi: refactor. break abi.Match interface.
In preparation for a new abi code generator, we are going to have Match
return a tuple item instead of a map of items. The map of items is
inherently flawed since it only does a good job of accessing top level
inputs. Events with nested data structures were not name-accessible
from the map.

To solve the accessibility problem, we will introduce code generation
that will act as a thin layer around the Item so that accessing data
doesn't require indexing (eg item.At(XXX) and so forth).