abi: fix for encoding/decoding of nested, static types (#65)
There was a critical flaw in the encoding/decoding functions that
mishandled nested static types. Prior to this commit, the functions
would use dynamic offsets for any type that wasn't a simple static type.
This is incorrect since the ABI spec defines the following:

	Definition: The following types are called “dynamic”:

	- bytes
	- string
	- T[] for any T
	- T[k] for any dynamic T and any k >= 0
	- (T1,...,Tk) if Ti is dynamic for some 1 <= i <= k

	All other types are called “static”.

In light of this fact, I added the Static method to abit.Type that
follows the rules outlined in the ABI spec.

Subsequently, as we decode static types, we need to keep track of how
many bytes we are reading --which is dependent on the type being
decoded. So I added a Size method to abit.Type that will compute the
size if the Type is static.

Finally, I addressed a critical shortcoming in abi's testing corpus by
adding explicit encoding tests.