abi: bugfix for fixed length list decoding (#67)
This commit adds support for decoding fixed length lists. The solution
also improves the architecture of the code as it leverages tuple
decoding instead of doing it's own decoding. This means we can remove
the Size method on abit.Type. After all, the ABI spec instructs
implementations to treat lists as tuples.