abi: bugfix. decode dynamic sized list with more than one element
There was a regression in #67 where we stopped checking the size
of an item after decoding it. This accumulated size is needed
when decoding nested tuples.

The Decode test was improved as to expose the particular bug.

Finally, the Encode test was simplified to make it easier to understand
how a nested tuple is encoded.