abi: return number of bytes read from abi.Decode
It is not uncommon for ethereum calldata to include an abi encoded
payload with extra bytes appended. In order to seperate the abi encoded
data from the extra bytes, Decode will return the number of bytes read
so that callers can safely extract the extra bytes.

This commit adds a test to ensure that Decode will work with input that
contains extra bytes.