genabi: improve names of generated structs
There are some major, breaking changes in this commit. Namely

1. Structs related to events are suffixed with Event. Eg Transfer
   becomes TransferEvent.
2. Structs which are nested in Event, Request, or Response structs are
   prefixed with the outer struct's name. Eg Offer becomes
   OrderFulfilledOffer. Or Order becomes CancelRequestOrder.
3. Descriptors (ie Event, Function) can reuse names so long as the
   arguments are unique. In this case, we suffix a duplicate descriptor
   with a sequential integer. The first occurrence is not suffixed and
   the 2nd occurrence begins with suffix 2.

Additionally there were some refactorings to simplify the, already
complex, genabi code. Specifically the goType and itemFunc functions no
longer take a Field struct but instead take a Type and Name string. The
idea here is to not depend too much on Field or Descriptor and instead
transform the data in the template.txt file where it is most relevant.