rlp: use values instead of pointers (#106)
I don't like using sync.Pool as a band aid for performance. The extra
complexity of pointers and the opaque nature of sync.Pool
makes reasoning about GC performance more difficult than it ought to be.

In the particular case of RLP decoding, the actual problem is that we
blindly decode nested structures. This wholesale decoding leads
more allocations than may be necessary. Instead, there should be an
API that allows callers to decode one "level" at a time. This reduces
computation and allocations at the cost of a little more work on the
caller's behalf.

A forthcoming pull request will contain this new API.

Revert "rlp: use sync.Pool for Items"
This reverts commit d4a1fafd415f74032ecf91608ab3a4dc30f70b10.

Revert "rlp: use pointers to rlp Items"
This reverts commit d9c1dbef59b1bec1ef6e36bbca03b64bc34cdca1.