rlp: alloc-free scanner
This commit makes a big change to how RLP data is encoded and decoded.
Instead of using an Item struct, the API now works directly with []byte
and [][]byte slices. The API no longer decodes arbitrarily nested data
structures and instead relies on the caller to successively create
Iterators. This allows callers to optimize their usage as to reduce
memory and computation utilization.