g2pg: improve reorg / tx handling. expose end block
The primary motivation for this change is to make it easier to develop
g2pg. Things like running locally, QA testing, and performance profiling
ideally require no steps other than: go run ./cmd/g2pg

In particular, I am wanting to introduce a -profile flag that will index
a set of blocks (eg 16,000,000 through 16,500,000) while taking a
comprehensive pprof profile and then once the blocks have been indexed,
the profile should be loaded in the browser. This kind of one-click
profiling will help myself and others make sure g2pg is fast and
accurate.

This change also fixed something that was bothering me about the reorg
tx handler. Now, if g2pg encounters a reorg of 10 blocks, it will
rollback the PG transaction such that no data was actually deleted and
finally the program will crash. If there is actually a reorg of this
magnitude, the operator will have to use PSQL to purge the blocks.