gethdb: update Hash to work with non-db blocks
The 'db' only knows about recent hashes so we use the Headers table in
the freezer to compute a hash for a particular block number.

There is a hashes table in the freezer, however, this table isn't
compressed and currently the freezer implementation assumes that it
works with compressed data. Since we don't use uncompressed freezer data
other than at g2pg startup, it didn't seem like a worthwhile change to
make in the freezer's hot path. That code gets called a lot.