cmd/g2pg: update profile access api
It should have occurred to me sooner that you need to source code to be
co-located to the invocation of 'go tool pprof' in order for the ui to
be fully functional. In light of this, it doesn't make sense for the
binary to start the pprof ui. The workflow now is to run the following:

	% ./g2pg -reset -begin 16000002 -end 16100000 -profile cpu

And then on a machine that has access to g2pg and the source:

	% go tool pprof --http=:8547 'http://g2pg:8546/debug/pprof/profile'

This will open up a browser and load the profile for the entire batch of
blocks that g2pg processed.