.github: disable concurrent testing
There is currently an issue with https://github.com/bmizerany/pqx
where multiple tests are trying to install Postgres at the same time and
when this happens the [Binary function][1] returns an os level error.

There needs to be some sort of locking in pqx but for now we will simply
run tests sequentially to avoid the issue.

[1]: https://github.com/bmizerany/pqx/blob/main/internal/fetch/fetch.go#L34