g2pg: refactor tx hash cache
I moved Transactions.Reset and Receipts.Reset to the place where we
setup our batch. This makes it explicit that the batch will not contain
leftover transactions or receipts. This is important because we plan to
skip the loading of certain blocks based on a bloom filter (coming in a
future commit).

After making the change to reset Transactions on each batch, I noticed
that iterating over each transaction became expensive. Instead, I reset
the hash cache only when we are unmarshaling a transaction and the
transaction was recycled --since this is the precise moment we want to
invalidate the cache.