e2pg: s/driver/task
Driver never really made sense to me. As we move the codebase to a place
where there are multiple chains and multiple concurrent indexing
operations, it became urgent to find a better name. I considered
"index" or "job" but ended up enjoying "task" the most. For example:

> To index a new chain, create a new task and configure it with the chain id, integrations, and start/end block.

> E2PG can run multiple tasks concurrently.

> To backfill a particular contract, create a new, temporary task and
configure it with the contract integation and the start/stop block.