cmd/e2pg: use pg tx by default (#146)
The reason to not use a DB transaction is that it adds significant
overhead and this overhead slows down large backfill operations.
However, not using transaction means that when e2pg is not crash safe.

A better long term solution may be to use transactions when the
processing batch is small, meaning it is incrementally adding blocks,
and avoid transactions when the batch is at max capacity, meaning we are
backfilling. This way there is no configuration. Another solution may be
to figure out how to avoid transactions altogether while maintaining a
crash only design.

For now, we will always use a db tx unless specified.