rlps: client sends hash to avoid cached reorg blocks
As of today the CDN doesn't cache data since it's likely to serve
invalid data in the case of a reorg.

The scenario:

	E2PG has block {n:1 hash:a parent:z} and then requests block {n:2}.
	E2PG requests {n:2}
	E2PG stores block {n:2 hash:b parent:a}
	Eth reorgs and {n:2} becomes {n:2 hash:c parent:a}
	E2PG deletes {n:2 hash:b parent:a}
	E2PG requests {n:2}

At this point the CDN will serve: `{n:2 hash:b parent:a}`

The fix:

	E2PG has block {n:1 hash:a parent:z} and then requests block {n:2}.
	E2PG requests hash for {n:2}
	E2PG requests {n:2 hash:b}
	E2PG stores block {n:2 hash:b parent:a}
	Eth reorgs and {n:2} becomes {n:2 hash:c parent:a}
	E2PG deletes {n:2 hash:b parent:a}
	E2PG requests hash for {n:2}
	E2PG requests {n:2 hash:c}
	E2PG stores block {n:2 hash:c parent:a}

The RLPS server code will ignore this value. It's only used as a cache
key for the CDN. Maybe later we can use this value to futher optimize
the server read path.

Eventually I need to enable caching on the CDN. For clients who aren't
running this code, and in the case or a reorg, their clients will be
stuck until they update to this commit. In other words: BREAKING CHANGE