e2pg: bugfix. crash when eth hashes aren't present
This is a sad UI bug. The system can be in such a state where these
slices are uninitialized and therefore [:4] will result in a panic.

The whole reason to [:4] is to make the hashes not take up so much space
in the UI and in the logs. So a better, safer way to deal with this is
to let printf deal with the sizing.

Closes #145