e2pg: use block number for reorg instead of hash (#153)
When a task indexes a batch of blocks it records the highest block
number in the task table. If a block within a batch is removed via reorg
then it is possible for integrations to contain rows relating to a block
that isn't present in the task table. In this case the integrations rows
are never removed.

To address this problem, the integrations will now receive a block
height that is to be removed and the integrations will delete all rows
above and including that block height.