e2pg: bugfix. tx with len(tx.S) == 31
The public key recovery code assumed that S would always be 32 bytes but
that's not the case. The fix was to make the sig to be exactly 65 bytes
and then fill in as mush as R and S as is available. This ensures that
we always pass 65 bytes to the recovery function.