e2pg: use pg_advisory_lock to limit 1 process per task (#159)
Deploys, or operational errors, can lead to multiple E2PG processes
running at a single time. This causes duplicated work but because of the
databases unique constraints there are no corruption issues. However,
it is best to avoid these duplication errors as it can be confusing to
deployment tooling and its operator.

Fixes #151