e2pg: move task locking closer to the work
I realized that there were potential issues with the prior locking
system. The lock was being released when the session was terminated.
However, the lock was acquired using a pooled connection. Meaning if the
pooled connection was closed, and the task was still being processed,
then another process could acquire the lock -- thus defeating the
purpose of the system entirely.

Instead, I changed the lock to be a transaction based lock. This means
that locking will only work when DB transactions are enabled. Which is
the default. This means that locks are more fine grained which has the
consequence that if multiple processes are running, the tasks will be
executed by 1 of the N processes and the 1 is determined by the first to
acquire the lock.