cmd/e2pg: don't exit on errors. sleep and retry forever
This is a big behavioral change in the way that the e2pg daemon works.

Initially I was certain that crashing is the best strategy. There is
something nice about a crash in that it will clean the pipes. It's good
for a process to be restarted every now and then. Additionally it forces
me to ensure that e2pg remains "crash only."

That being said, this model relies on tools like PAAS schedules or
Systemd to keep the overall system online in the event of an error. In
practice this places a lot of the reliability burden on the e2pg operator.

Instead, I would like to try not crashing on errors but instead just
sleeping for a second and then retrying the task.

I did take some time to look at the data structures in the Task.Converge
method to ensure that there isn't any consistency issues to deal with. I
didn't find any major issues. We call Reset on the block and it's nested
structs prior to marshaling data. So we should be good. The only other
concern would be the geth.Buffer but I think we are good there as well.

I would like to get rid of geth.Buffer soon.