e2pg: use stable task id (#173)
This is part of a larger change in E2PG. Instead of users defining
tasks, E2PG builds tasks based on user configured integrations.

Each chain (defined in the Ethereum Sources field in the config)
will have a main task. Integrations that are following the chain will be
processed in the main task. If an integration requires a backfill, a new
task will be created. Backfills are not a part of this commit and will
be added later. Backfills will need to coordinate to not overlap work
with the main task.

Given this archetecture, a task id can simple be: the chain id, a string
tag that is one of 'main' 'backfill' and an optional backfill string tag
that is unique amongst other backfills on the chain. For example:

1-main

1-backfill-foo