abi2: bugfix. tx-only processing and mismatched block/column names
The block name's need to match the list of available options:

https://github.com/indexsupply/x/tree/main/docs/e2pg#block

However, the datbase column name doesn't need to match those options so
long as the block's column value references the table's column name.

Instead of the abi2 code using the column name to look up the block
data, it should use the block object's name.