e2pg: refactor. s/node/src s/integration/dest
The old Integration interface in particular didn't capture the full
breadth of what an Integration can do. For the purpose of a task, it
only cares about inserting, deleting, and getting events for filtering.

I'm preparing the e2pg package so that i can move the code in the
e2pg/config package up a level.