e2pg: new task manager. move e2pg/config to e2pg
The config package was a hack to workaround import cycles. The prior few
commits have been to undo the import cycle fiasco.

The new task manager provides a way to restart tasks. This will be used
when we have an interface for adding/removing tasks without restarting
the process.

The task manager also loads tasks from the newly created database
tables. The config file can be used alongside the database and the
database config will be overwritten by the config file in the case of
conflicts.