e2pg: rebuild dashboard (#181)
The new dashboard forsakes the old way of delivering task updates.
Prior to this commit, we showed a dashboard based on in memory
statistics. This design only worked for a single process. When you
are running multiple processes for fast backfilling the dashboard
would no longer work.

This commit stores task/integration status data in the database on the
existing task table. I want to keep the task table small since it is
a mission critical table. But I think the multi process benefit
justifies the cost of the additional columns.

Additionally, the dashboard now will show which integrations are running
on a particular task. You can press the 'i' key on the dashboard to toggle
the display of each task's running integrations.