e2pg: remove task id
This change is needed for the new backfill model. The goal is to remove
the notion of creating tasks from E2PG's API. Instead, E2PG will use the
config to determine which tasks need to be created. Each ETH source will
have a main task and a backfill task. A task will be identified by it's
src_name and backfill columns.