e2pg: add integration update table (#188)
This table will be used to control backfill tasks. The task table keeps
a history of blocks processed, but they are used for detecting a
re-organization. It does not track an integration's individual progress.

By introducing the intg table, we now can know the precise set of blocks
that an integration has processed.

This will enable a backfill task to group a set of integrations that
need backfilling and process block history based on the needs of the
integration.

This commit also introduces the notion of e2pg.{task,intg} pruning. We
keep around the last 200 blocks for each src_name / src_name,name. This
should help keep queries on this table fast while also preserving
history for reorgs and debugging.