e2pg: adjust intg task pruning
After reading the config from the file and database, E2PG will insert an
intg backfill row for the config's requested start block.

To determine if the backfill has been complete, we will take the highest
block number for the integration and compare it with the lowest block
number on the main task.

This means that we only need the smallest and largest block number for
each integration / source in the intg table.