e2pg: introduce a destRange for backfill filtering
This is needed because there will soon be a backfill task per source
that will write to destinations that may or may not have the same level
of progress. E.g. it is likely that there is an integration that has
been backfilled from block: 0, 100 but on e2pg restart, a new
integration has been added that requires a backfill of 0,200 and so the
backfill task will work on both integrations but starting at 0 again. So
the first integration will not be asked to insert data until the
backfill task reaches 100 again. Similarly, an integration may not accept
new Insert calls if it is already up to date with its main task while
other backfill tasks still have blocks to process.