e2pg: bugfix. only write an intg update if an integration has changed
Now that loadinsert is filtering data, it is possible for there to be no
change in the intgUpdate data. This will cause old data to be copied
into the intg table when the write function is called. The write
function will fail because there is a database constraint that prevents
duplicate data in the intg table.

The fix for this is to only write an update for an intg if it has
changed. This avoids the db constraint error and lets e2pg go about its
business.

Another commit should remove an intg from a task once it no longer will
be updated. In other words, once a backfill is complete, we should stop
processing it.