abi2: bugfix. not matching events based on number of indexed fields
I noticed that certain abi events were not being saved to my table and
upon further inspection I realized that when I had setup my integration
I only selected 1 out of 3 of the indexed fields on the event.

When an integration is testing a log to see if it should index it, it
uses two methods:

1. topics[0] byte comparison
2. the number of non-empty remaining topics matches number of indexed
   inputs in the abi

In my case, the 2nd method was failing because of a bug. The code was
reporting the number of indexed inputs based upon the selected inputs.
Instead, we should simply look at all the top level inputs to see if
Indexed=true.