abi2: validate integration. add required columns
In order for the abi2 package to properly delete
data during a reorg, we needed to make sure that we have
intg_name, src_name, and block_num columns available.

Additionally, this code sanitizes user input for table and column names
since we are building sql strings without bind parameters.