e2pg: improve task manager setup
Having the loadTasks and task.Setup function execute within the task
manager's mutex is nice because it makes it easy to reason about order
of events. In particular, task setup may insert into the intg table for
backfill tasks. If there is already a backfill task underway, we are
changing the intg table under the feet of the currently running backfill
task. It's much simpler to assume that a single writer is writing to the
intg table on behalf of a src/task/backfill combination.