e2pg,jrpc2,rlps: remove geth.Buffer
This is 1 step back in order to make a couple of steps forward. The goal
of the geth.Buffer was to hold onto arrays so that concurrent task
processing could take advantage of already allocated memory. However, I
do not like this design. Instead of having a single client for reading
data from the file system, I would like to heave each task partition to
have its own client with its own buffer. This way, we don't have to
manage a shared resource.

The 2 steps forward will come in the next couple of commits. Each task
with have a []part field. The part will hold onto a subslice of blocks
to process and it will also have a source field. This means that if we
have 10 concurrent parts in a task, each part can use its own Source
which mean the Source can deal with memory management and not worry
about resource sharing.