e2pg: bugfix. {main,backfill} task both start at block 1
If you setup an integration where the eth_source defines a start block
of 1 and your integration sources also start their backfill block as 1
you will hit this bug where the backfill and main task are running out
of sync. Which means you will get unique constraint violations.

The bug is that the task.Converge function would check to see if a task
is complete by first seeing if the task.stop value is greater than 0
(implicitly indicating that it had been set) and then checking the stop
value against the task's local block height. In our case, the main task
setup added a starting block of 0 and began processing later blocks. The
backfill task loaded it's start / stop values based on where the main
task was at, namely, the backfill task loaded a stop value of 0 since
that is where the main task started. But because the stop value was
zero, and not just un-initialized zero, the check to stop the task was
never evaluated.

This fix removes the implicit stop > 0 check and instead checks to see
if the task is a backfill task. It is assumed that a main task never
stops. This may be incorrect someday, but we can cross that bridge
later. For now, if a task is a backfill task, and it has a stop value of
0, (or of the local block height is greater than or equal to the stop
value) then the task is stopped with error: ErrDone.

The resulting behaviour is that if you set a global start block to the
same value as an integration stop block then the backfill task will
start and then immediately stop.