e2pg: refactor. move integrations/destinations into task.part
This commit is a follow on to [1] in which we introduced a 'part'.

When processing a Task's batch of blocks, we concurrently use the task's
source (for downloading block data) and the task's destination (for
decoding block data and writing to pg) and since the source and
destination expect the caller to manage concurrency, we must use them in
a part.

The motivating idea behind all of this is to encapsulate memory
management for downloading and decoding. This way the jrpc2/rlps clients
can maintain their own buffer to reuse between requests. Likewise, the
abi2 package will decode ABI data using a recycled buffer. The calling
code in e2pg.Task doesn't care about how downloading and decoding are
reusing memory. It just needs to keep a client for each concurrent go
routine.

[1]: 71cfc671ab67177080793a5c1e748299ad603b94