e2pg: s/intg/ig (#199)
I go back and forth on var names. I tend to like short names for small
functions and non-abbreviated names for global concepts. However,
integration is a very long word. It's a PITA when grepping logs of
writing sql queries. To that end, I'm going to use IG for exported,
abbreviated references to Integration and ig for un-exported,
abbreviated references.