jrpc2: return rpc errors on hash/latest calls
See previous commit. I was confused as to how a particular Shovel
instance was "ahead" but now I see. When the RPC call to Latest returns
a JSON RPC Error then the latest will be 0. Thus the local node is
ahead. We were swallowing the error. Now we are not.