shovel/config: refactor config into new package (#218)
In addition to moving code out of shovel/task and into shovel/config,
this commit also changes how we setup the database for the configured
integrations. As soon as the program starts up, it reads the json
config into a config.Root struct. It then will validate all the
input and check the config.Root for referential integrity (ie all
your columns are specified and things match up). It will then migrate
each table defined in the config. If a table defined in the config
has new columns (this can happen when a new integration reuses an
existing table but define new columns) then those columns
are added to the table in the database.