shovel: bugfix. address filter for getLogs
when shovel creates a task, it starts with an empty Filter and then
merges each integration's filter into it's initial filter. Since the
empty filter contains no addresses, and given that if an integration in
a task doesn't specify an address then there can be no address filter,
then it was impossible to use an address for filtering.

This fix solves the problem by treating empty filters in a special case.
If the filter is empty, and if you merge another filter with an address
list defined, then the empty filter will take on the merged filter's
addresses.