shovel: refactor. db advisory locks
pg_advisory_lock expects an int64 value as the lock id.

wpg.LockHash aims to be a nice way to manage how we map domain
locks onto the int64 range. To do this we use the Fowler–Noll–Vo
hash function (fnv-1a 32). The idea behind using the uint32 variety is
that we have to cast to an int64 and since our domain is small (on the
order of 100s) it seemed like the safest thing to do is go from uint32
-> int64. As an added precaution, I added a map to check for collisions.
This made the function expensive, but as you can see in this commit we
do all the computation at startup.