shovel: bugfix. backfill unable to complete last block
There was a bug when checking boundary conditions
on a backfill/latest task that caused the backfill to not complete and
in some cases the backfill would attempt to insert a duplicate block.
Thankfully the unique index prevented the duplicate data from being
added but it prevented the backfill from completing its final block.

Latest block processing is not affected by this bug.

The fix was to align the assumptions in the data model with the code. In
particular, the max(ig_updates.num) column was assumed to be the last
block processed -- which means that the next block to process should
start at max(ig_updates.num) + 1. The code was not doing this.

I have corrected the code and added more code comments to help myself
better understand this system in the future.

However, i would like to burn the current design to the ground with
fire. It is too complicated and error prone. I have a few ideas on how
to make things much simpler and will make these changes in the future.