shovel/dig: fix bug when indexing uint256[]
The dbtype change was getting hung up on the type being uint256[]. The
brackets were causing the bit calculation to fail. This was an un-proven
optimization anyways, so it seems safe to always use the uint256 package
to decode the number if the number type is a uint*.

I do wich that the value passed to dbytpe was a scalar type and not an
array. I will address that later.

The clear call was something that I noticed while looking at this code.
Since we reuse the rows, we need to clear the data. This may have casued
incorrect data to be written.