shovel/task: simplify task model
This commit introduces a big changes to how Shovel works:

Previously there was a main task and a backfill task per ethereum
source. The task would download a batch of blocks and then distribute
those blocks to the configured integrations. The main task would
start right away, and depending on the integration's source start
field, spawn a backfill task. The backfill tasks would be grouped
by source and as the backfill task progressed the integratiosn would
be called if they needed the data (tracked in the ig_updates table).

This was overly complex and it caused me anguish.

This new system is much simpler. One task per integration/source
combination. If a task needs backfill, it starts at the `start`
value and continues as the chain progresses. Ethereum source data
access is conserved through a cache introduced in the jrpc2 package.