shovel/jrpc2: bugfix. validate rpc resp before caching
There was a bug report where the task was processing block >
19000000 and the task attempted to insert num=0 into the task_updates
table. The bug was observed when it attempted to insert num=0 for a
second time and thus causing a unique constraint violation.

To insert num=0 into the task updates table, the last block in the
task's processing loop must have had num=0. This could either be a
zero-value issue or the call to jrpc2.Get was returning blocks with 0 as
a header number. I didn't see anything in the code to indicate that the
former is a possibility. Since the bug report came from a free node RPC
provider, and this provider has had some issues in the past, I am
wondering if it was returning invalid data for a eth_getBlockByNumber
request. This commit addresses this possibility by doing a validation
check on the eth_getBlockByNumber response. If it doesn't match the
request then an error is returned.

But that's not all. In this particular bug report, we saw that unique
constraint violation errors went away upon restart. This seems to
indicate that the invalid data was being cached. This commit addresses
that issue to since invalid data will never be written to the cache in
the first place.

Finally, in the loadinsert method, as we are concurrently processing
blocks we are keeping a pointer to the first and last block in the
batch. This was not protected by a mutex even though it ought to be. I'm
not sure if this was causing problems, but it definitely stood out as a
potential issue. So I added a mutex.