shovel/dig: bugfix. reset data for events w/ null trailing dynamic
A caching bug. When an event has a null trailing dynamic
field is processed after an event with a non-null trailing
dynamic field and the events are in the same batch then
it is possible for the data to be incorrect.

The prior code was correctly resetting the result cache for the variadic
portion of the result but neglected to reset the single portion of the
result.