shovel/web: bugfix. localhost dashboard access
This code incorrectly assumed that if you were accessing the dashboard
over localhost then you were running shovel in dev and would have access
to the html templates on the fs. This is incorrect. You may want to
access the dashboard over localhost using the released binaries, in
which case you don't have the html on the fs. So now we always fall back
to the html in the compiled binary. We also log in case something is
going wrong in dev.