shovel/dig: refactor. adjust types for uint256 values
This should change no behaviour. Instead of converting a Decimal value,
we will simply return the uint256 and let pgx call:

	Value() (driver.Value, error)

on the uint256.Int. I am making this change so that both:

	func (lwc *logWithCtx) get(name string) any
and

	func dbtype(abitype string, d []byte) any

will not strip away the uint256 type. This will be helpful when I change
the way that filter.Accept works so that we can do numerical filtering.