shovel: index traces (#249)
This is an early draft at indexing traces. There might be more work
to do here but I am getting this out for a particular user who needs the
data. Luckily there wasn't much code required. The proceeding commits
where we added the ability to filter on numerical and string values make
this feature more reasonable because the trace data can be large.