shovel/jrpc2: fix mem corruption on http error for trace indexing
The block, and consequently a block's transactions are cached for a
period of time. When a shovel task is batching index operations, it may
process block=i (caching txs/traces in memory) and if block=i+1 received
an error, the database transaction would be aborted and the batch of
transactions would be processed again. But because block=i's txs/traces
are cached, and because of the bug that this commit fixed, the jrpc2
client would download block=i's txs/traces again -- thus duplicating the
data. The batch of blocks would fail because the shovel task would be
attempting to insert duplicate data.

The postgres unique index prevented the duplicate data from being saved
in the database, but the task would be unable to make progress until the
cached data was deleted.

This fix will reset a transaction's traces for each request thus
eliminating the possibility for duplicate data.