shovel/dig: add type conv. prior to filter checking
related: 42721b26a79509c3847f5f9ad1b8e5adbaf01391

In the related commit, we added the ability to do meaningful comparisons
on typed numbers, bytes, strings, etc. This checks work great for data
that comes from logWithContext.get() because that method returns decoded
data. However, in order for these filter checks to work on all cases, we
need to deal with abi encoded data. Namely, we need to decode it into a
Go type, cast it to any, and then hand it off for filter checking.